+++ Sources/VideoKit/VideoPlayer.swift
@@ -25,24 +25,64 @@ import SwiftUI
 /// ``SwiftUICore/View/videoPlayerConfiguration(_:)``.
 public struct VideoPlayer: UIViewControllerRepresentable {
 
+    /// Create a video player.
+    ///
+    /// - Parameters:
+    ///   - videoURL: The video URL to play.
+    ///   - time: An external player time binding, if any.
+    ///   - configuration: The player configuration to apply, by default ``VideoPlayerConfiguration/standard``.
+    ///   - controllerConfiguration: The controller configuration to apply, if any.
+    ///   - didPlayToEndAction: An action to trigger when the player reaches the end, if any.
     public init(
         videoURL: URL?,
         time: Binding<Double>? = nil,
+        configuration: VideoPlayerConfiguration = .standard,
+        controllerConfiguration: @escaping (VideoPlayerController) -> () = { _ in },
         didPlayToEndAction: Action? = nil
     ) {
         self.videoURL = videoURL
         self.time = time
+        self.configuration = configuration
+        self.controllerConfiguration = controllerConfiguration
         self.didPlayToEndAction = didPlayToEndAction
     }
 
-    @Environment(\.videoPlayerConfiguration) var configuration
     @Environment(\.dismiss) var dismiss
 
     public typealias Action = () -> Void
 
     private let videoURL: URL?
     private let time: Binding<Double>?
+    private let configuration: VideoPlayerConfiguration
+    private let controllerConfiguration: (VideoPlayerController) -> ()
     private let didPlayToEndAction: Action?
+}
+
+/// This type can be used to configure a ``VideoPlayer``.
+public struct VideoPlayerConfiguration: Sendable {
+
+    /// Create a custom video player configuration.
+    ///
+    /// - Parameters:
+    ///   - autoPlay: Whether to auto-play video when the player appears, by default `true`.
+    public init(
+        autoPlay: Bool = true
+    ) {
+        self.autoPlay = autoPlay
+    }
+
+    /// Whether to auto-play video when the player appears.
+    public var autoPlay: Bool
+}
+
+public extension VideoPlayerConfiguration {
+
+    /// The standard video player configuration.
+    static let standard = VideoPlayerConfiguration()
+
+    /// A list-specific video player configuration with auto
+    /// play disabled.
+    static let list = VideoPlayerConfiguration(autoPlay: false)
 }
 
 public extension VideoPlayer {
@@ -55,9 +95,11 @@ public extension VideoPlayer {
         controller.modalPresentationStyle = .fullScreen
         controller.playerTime = time
         controller.playerDidPlayToEndAction = didPlayToEndAction ?? {}
-        configuration(controller)
+        controllerConfiguration(controller)
         controller.player = createPlayer()
-        controller.player?.play()
+        if configuration.autoPlay {
+            controller.player?.play()
+        }
         return controller
     }
 
@@ -88,22 +130,6 @@ private extension VideoPlayer {
         let item = AVPlayerItem(asset: asset)
         let player = AVPlayer(playerItem: item)
         return player
-    }
-}
-
-public extension EnvironmentValues {
-
-    /// This value can be used to configure a ``VideoPlayer``.
-    @Entry var videoPlayerConfiguration: (VideoPlayerController) -> Void = { _ in }
-}
-
-public extension View {
-
-    /// Apply a custom ``VideoPlayer`` configuration.
-    func videoPlayerConfiguration(
-        _ config: @escaping (VideoPlayerController) -> Void
-    ) -> some View {
-        self.environment(\.videoPlayerConfiguration, config)
     }
 }
 
@@ -147,10 +173,13 @@ public extension View {
     struct ContentView: View {
 
         var body: some View {
-            VideoPlayer(videoURL: VideoPlayer.sampleVideoURL)
-                .videoPlayerConfiguration { controller in
+            VideoPlayer(
+                videoURL: VideoPlayer.sampleVideoURL,
+                configuration: .init(autoPlay: false),
+                controllerConfiguration: { controller in
                     controller.showsPlaybackControls = false
                 }
+            )
         }
     }